requirements.txt
asgiref==3.7.2 ; python_version >= "3.12" and python_version < "4.0"
crispy-bootstrap5==2023.10 ; python_version >= "3.12" and python_version < "4.0"
django-crispy-forms==2.1 ; python_version >= "3.12" and python_version < "4.0"
django-phonenumber-field[phonenumbers]==7.2.0 ; python_version >= "3.12" and python_version < "4.0"
django==4.2.7 ; python_version >= "3.12" and python_version < "4.0"
environs==9.5.0 ; python_version >= "3.12" and python_version < "4.0"
faker==20.0.0 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.20.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.12" and python_version < "4.0"
phonenumbers==8.13.24 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.12" and python_version < "4.0"
transliterate==1.10.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
